Fix some tests
