+++ Sources/TuistCoreTesting/Extensions/AargumentParser+TestData.swift
@@ -1,7 +1,7 @@
 import Foundation
 @testable import SPMUtility
 
-extension ArgumentParser {
+public extension ArgumentParser {
     static func test(usage: String = "test",
                      overview: String = "overview") -> ArgumentParser {
         return ArgumentParser(usage: usage, overview: overview)
